+++ packages/mirai_framework/lib/mirai_framework.dart
@@ -1,4 +1,0 @@
-library mirai_framework;
-
-export 'package:mirai_framework/src/mirai_action_parser.dart';
-export 'package:mirai_framework/src/mirai_parser.dart';