+++ packages/stac_framework/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ## 0.0.1
 
-* Init Mirai Framework package
+* Init Framework package